add filter support to allpaths, somepath, rdeps
Summary:
The `somepath` and `allpaths` queries can help answer interesting dependency questions like "does this target depend on a library", but it can return misleading results if the dependency chain looks like:
```
  root_target -> embedded_binary
  embedded_binary -> lib1
  root_target-> lib2
```

Both `somepath(root_target, lib1)` and `somepath(root_target, lib2)` return a path in this example, but I may want to know if the dependency on a library occurs because root_target uses it, or some other binary.

With this, you should be able to do something like:
`somepath(root_target, lib1, target_deps() - kind('.*binary.*', target-deps())`

And discover that `root_target` does not "directly depend" on lib1.

Reviewed By: cjhopman

Differential Revision: D64610211

fbshipit-source-id: d27b78adbb7704733d6e27734b34cd3cc64df368