Return Err if cannot downcast the value to a dict in select construction
Summary:
We get a few panics due to trying to unwrap a `None`. The issue is caused by having nested selects which should result in an error. This diff converts soft_error that was thrown before to hard error so that consumers don't have to deal with the errors downstream.

Left a TODO for adding a type restriction.

Reviewed By: blackm00n

Differential Revision: D64835697

fbshipit-source-id: 6fc6a34f84094d9b31e543f60f351924b343bcc2