refactor linker extra outputs to split artifact declaration from flag generation
Summary:
In order to support the linker extra outputs flow with distributed LTO we need to separate the output declaration from binding during the flag generation step. This will allow the dynamic output actions of the distributed thin LTO to bind to the correct outputs.

This also attempts to unify the flow for cxx_binary and cxx_library so that the output providers are always set in the cxx layer.

Reviewed By: NuriAmari

Differential Revision: D64992811

fbshipit-source-id: 35c58546f7e8c3913fe35270852976d3b08ff7a0