Add option to write ExitResult errors into a file
Summary:
Adding a flag that allows ExitResult errors to be written to a file. This file include more client-side errors than invocation record. Implemented to help CI with error clustering.

This currently has 2 problems:
1. It doesn't include super early errors that don't even make it to command execution (This is difficult fully cover and my assumption is that those super early errors are rare)
2. If an error occurs after a build execution is complete (ex. Failed to materialize, out of disk space, etc.). Build errors would not be written and it will be replaced with the late error. This is not a trivial fix due to our current design.

Ideally this (And maybe build reports too) should be merged into invocation records so we have a single log with everything, and it all makes it to scuba

Reviewed By: christolliday

Differential Revision: D64728650

fbshipit-source-id: 95d4c4ec0f7452cd3506da2292b5d4bdd5977e6c