Add link for the types in the type signature in the doc
Summary:
# This Diff
This diff passes the `TypeRenderConfig` across all different render functions.

If given `TypeRenderConfig::Default` it will render the output as before. If given `TypeRenderConfig::LinkedType`, it will render it with linked type.
 ---

In markdown, we cannot have the link for the "```" code block. i.e.
```
def get_artifact() -> [Artifact](/path/to/artifact)
```
Above linke will not work.

We need to use the html block here, e.g.
```
 <pre class="language-python">
  <code>
    def soome_function() -> <a href="/path/to/type">Artifact</a>
  </code>
</pre>
```
----

We can click these type to it's doc.
 {F1947337749}

Reviewed By: JakobDegen

Differential Revision: D64875220

fbshipit-source-id: 9ea748919e5815b89bea294e9cc6777a81b86e56