Refactor display for all kinds of Ty
Summary:
# Context
This diff refactor display for all kinds of Ty that allow it to render according to a config which is used for  render linked type in later diff.

# This diff

This diff refactor the display for all kind of Ty.
It adds this method for these types which takes another config used for render

```rust
pub(crate) fn fmt_with_config(
        &self,
        f: &mut fmt::Formatter<'_>,
        config: &TypeRenderConfig,
)
```

The config definition is here
```
/// Configuration for rendering types.
pub enum TypeRenderConfig {
    /// Uses the default rendering configuration.
    Default,
    /// Uses for linked type in doc
    LinkedType {
        /// The map from type to it's path link
        ty_to_path_map: HashMap<Ty, String>,
    },
}
```
For convenience, it also adds type like `TyDisplay` if needed which used to implement `Display` and `Ty` can have method `.display_with(config)` to return this type for format.
```
pub(crate) struct TyDisplay<'a> {
    ty: &'a Ty,
    config: &'a TypeRenderConfig,
}
```

Reviewed By: JakobDegen

Differential Revision: D65174168

fbshipit-source-id: 233ce88d043301fba8b49e15dc98dd10a2beaf26