Provide type to path map for the linkable type signature in doc
Summary:
# Context
This diff is used for the linkable type in the type signature in the doc.

# This diff
Since we are implementing the linkable type in the type signature, we need know the information about what path should we use for the link. As a result, we need to pass a map from the type to the path to the render process.
The map can be generated from all the `DocModule`s
So this diff adds a type `DocsRender` which is generated from all the top level `DocModule`s. The filed `ty_to_path_map: HashMap<Ty, String>`  in `DocsRender`  is what we need here.

----
This diff changes some generated markdowns' path. They are belong to this case:
For `DocModule` (e.g. the module name is foo) who have no functions or properties definitions, we will generate markdown with such path:
* Before: `base_path/foo.md`
* After: `base_path/foo/index.md`
However, `docusaurus` will render these two cases same and the html path will also the same.

e.g.
* Before:
 {F1947328936}
* After:
 {F1947329099}

Reviewed By: JakobDegen

Differential Revision: D65004622

fbshipit-source-id: d6f13b061a4417a359eac14cc3a47e55d5c9aed6