Include error tier when ranking errors
Summary:
For errors with no tags we currently ignore the tier when determining the best error. This is fine when using the best error to determine the best tag, but it's a problem if we are trying to determine the tier of the best error.

This should function as if the tier on an error was a tag ranked at the bottom of the list of tags for that tier, like this:

```
env tags
tier0 tags
'environment' tier
'tier0' tier
input tags
'input' tier
unspecified
no tier
```

Although the tier0 tier is ranked below the environment tier, tier0 and environment errors aren't offset so it's still possible for tags of each type to take precedence over the other.

This would be easier to reason about if we exclusively derived tiers from tags and removed the ability to set tier independently. This seems potentially worth doing later.

Reviewed By: Will-MingLun-Li

Differential Revision: D65092673

fbshipit-source-id: 45d3d30535fb8de55727b85b7fb7294b9c4de663