Add workaround for poor watchman performance
Summary:
I've identified a perf issue with watchman when using scmaware queries
and empty_on_fresh_instance=false. This is hit when changing the merge base commit
and it causes the fresh instance to be really expensive, much more so than
doing the same empty_on_fresh_instance=false but without a clock.

This works around it by always passing empty_on_fresh_instance=true if we have a
previous clock, and then on a fresh instance we send a new query with
empty_on_fresh_instance=false and no old clock. This gets fine performance.

Reviewed By: JakobDegen

Differential Revision: D65360614

fbshipit-source-id: ba67c316e68b07594e7089bc2cf177e93b478cbb