tests: RFC: Don't require spelling out the targets
Summary:
The target definitions for core tests are extremely repetitive, cannot be automated by autodeps, and so are a bit of a pain to manage. Instead, let's just do it automatically for the user

This is a bit of magic, but it reduces the barrier to writing new tests and makes it less likely for people to screw it up, so I think an improvement overall.

Will update the remaining directories if the direction looks good

Reviewed By: scottcao

Differential Revision: D64726383

fbshipit-source-id: 8df79c7ad977b75e2a19b0fcdb2d819f79246d29