Fix for unexpected behavior from precise closure captures
Summary:
we tend to do something like:

```
let mut panic = MaybePanicOnDrop{panic:true};
...
async move {
  ...
  panic.panic = false;
  ...
}
```

But due to what I'd call a bug in rust, this has confusing behavior and doesn't do what we intended. See this simplified example:
https://play.rust-lang.org/?version=nightly&mode=debug&edition=2021&gist=374938d510161faaf16b92a06089fc4c

Add a function on PanicOnDrop for setting the flag so that we capture the entire object.

Reviewed By: JakobDegen

Differential Revision: D65362039

fbshipit-source-id: 06fe4cd932ad723b831d0157cc0683598317e401