Add missing raw_headers attr in cxx_library.bzl
Summary:
This was missing, which wasn't a big deal before, because
the raw headers were added as artifacts elsewhere and the specifics
didn't matter here. But it matters when implementing
`raw_headers_as_headers_mode` which does require processing the raw_headers
into a private headers hmap. So just add this now to make things simpler
later.

Differential Revision: D65517242

fbshipit-source-id: 11d3b03999446dd0885078a3812ae5e58a25b446