Segregate auto-generated .bcfg files
Summary:
Right now the `update_overrides.py.` script preserves existing content in the `*_overrides.bcfg` files, meaning that it doesn't remove incorrect/stale overrides. To fix this:

* Treat generated target platforms and constraint values separate from the rest.
* Always generate these from scratch each time, overwriting existing contents.

To make the separate clear, the `*.bcfg` files are now structured as follows:

```
tools/buckconfigs/fbsource
├── platforms.bcfg        # Manually curated target platforms.
├── constraints.bcfg      # Manually curated constraint values.
└── generated
    ├── platforms.bcfg    # Auto-generated target platforms.
    └── constraints.bcfg  # Auto-generated constraint values.
```

NOTE: Passthrough constraint values are not being respected right now, but this is fixed in D65477361.

Reviewed By: mojsarn

Differential Revision: D65533663

fbshipit-source-id: 64ca3f64034882e68760d422ce9292f0c861e632