Improve link_group analysis time by 99%
Summary:
Before, we could see that analysis takes ~100 seconds.
https://www.internalfb.com/buck2/0c139e16-2877-4837-9d3c-b3cc2324ee61
{F1955455916}

https://www.internalfb.com/buck2/0d0e21b6-f3c3-4069-82e7-b09521affefc
Now, it takes ~1 second. {F1955455104}

# HOW??

Currently, link groups allow you to pass filters so that you can do something like:

```
(":DepPool", "tree", "build_pattern://anything/in/this/...")
```

and that raw string is parsed in [build_target_pattern.bzl](https://fburl.com/code/nxelsaxu) and compared against various nodes in the `LinkableGraph`. **This isn't inherently slow... it's that the syntax is rigid, and BizApp was using nearly 100 of these filters to get the nodes they wanted in the graph.**

Furthermore, they had many filters which looked like this:
```
(":DepPool", "tree", "build_pattern://path/to/a:target")
```

this is insanely wasteful as it's comparing that string against everything in the linkable_graph which hasn't been resolved yet. It could restated as:

```
("//path/to/a:target", "tree")
```

By fixing these two issues:
  - using a new custom rule to do all the filter in one go.
  - removing filters on explicit targets

we see a massive speed-up in perf.

Reviewed By: mogers

Differential Revision: D65607704

fbshipit-source-id: e232ea49d23931bc01b447237c959067eea00a51