completions: Higher level timeout
Summary:
Importantly, instead of timing out only the target completions, this times out the entire command, regardless of the details. Shell completions not hanging indefinitely is really important, so we need to be getting this right.

I had considered doing this even higher up at the shell level, but hopefully this is enough?

Reviewed By: perehonchuk

Differential Revision: D65657700

fbshipit-source-id: b87f03b894ad17c1a0b37871652a621826ad4214