Do not include live pars in standalone resources
Summary:
With inplace by default enabled if a python binary is added as a resource the link tree would included as a hidden_resource. This caused standalone builds to fail - since we do not allow hidden resources in standalone pars.

This diff checks the resources to see if we are including a python binary. If so we add the standalone output as resource.

We track default_resources so that when building an inplace par we can include resources as inplace.

https://fb.workplace.com/groups/python.builds/permalink/3888593271462086/

S461975

#buildmore

Reviewed By: christycylee

Differential Revision: D65561372

fbshipit-source-id: e964b8bcbe05241c9ff2326d7641c1ae11b54293