Refactor adding best error fields to invocation record
Summary:
Use the ProcessedErrorReport for both buck2_errors and the best_error_* columns in buck2_builds to de-duplicate some code, ensure we are logging the same data, and to simplify later diffs.

`ProcessedErrorReport.tier` is based on `buck2_error::get_tier()` which has a subtly different implementation to `ErrorReport::category()`.
We should use `category` instead of `tier` since it matches buck2_builds `error_category` and is calculated in buck, not ingress.

Reviewed By: iguridi

Differential Revision: D65770810

fbshipit-source-id: 902c811deb96e1b18647f67d8c2e521d0fb13193