Add target property to disable used class
Summary:
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

We had to disable the used classes plugin for incremental targets. See D65209518 for more information.

We have observed a spike in action execution frequency, which may be a consequence of disabling the plugin. We want to understand how much the disablement contributes to the spike. Our plan is to disable both incremental and used classes on 4 targets to see how the data changes.

### This diff
$title

Reviewed By: hick209

Differential Revision: D65671277

fbshipit-source-id: 597dd5a2e1546e30f8622364b66b1f6f67f347b8