Filter out targets from PTT with no typecheck sub target
Summary:
Maggie ran into an issue where a custom rule was creating a python-like library, which PTT picks up on. The issue is that this custom target doesn't have a "typecheck" sub-target, which means we try to type check it and fail due to it not existing.

This diff creates a new `do_ptt_filter()` function, which works the same way as `do_filter()`, but also filters out targets without a `typecheck` sub-target.

Reviewed By: grievejia

Differential Revision: D65831320

fbshipit-source-id: 3e79cbf5338c2f84c21944e86d6a51c101441161