Remove sharding run for PTT in batch.bxl
Summary:
Sharding didn't help improve type checking, and it adds extra computation to do the buck build graph for python files. It's not necessary anymore, so let's remove it.

This diff turns off sharding in the BXL script that actually runs type checking. The flag is not removed, since it's still used in some places and I don't want to risk breaking them.

Reviewed By: ljw1004

Differential Revision: D65886198

fbshipit-source-id: bd95c9fb5f3f801a0957a0a059b334a988424fd6