Private modulemap
Summary:
Currently, if you are building a `apple_library` with private headers (but you want them seen by local `.swift`) you have to store them in `exported_headers`.

Once your buck target sees `exported_headers` it has to generated a Bridging header and then regenerate a module map, and so it hurts build speed, and leaks abstraction.

This uses the `private` module map so that we can access private headers with `@_implementationOnly import Foo_Private` module.

By default, this we don't generate the private module. This is because it will now fail if _private_ headers have any cxx in them. As such, you have to opt-in on your target via: `enable_private_swift_module = True`.

Reviewed By: rmaz

Differential Revision: D65885148

fbshipit-source-id: 16a485b3f15cedb9a44711622b0485456dd6b8ed