Add tags to replace error tiers
Summary:
Motivation in D65688581

In order to always use tags to determine error tier, we can use error tags that only represent a tier to replace existing usage.

It seems slightly preferable to require more specific tags as in D65688576, D65688580. We could remove these later if we decide to enforce that.

Temporarily renaming enum names in error.proto::ErrorTier because:
- otherwise this would cause a name collision, apparently siblings enum values end up in the same namespace
- renaming enum value names shouldn't break binary protobuf compatibility
- protobuf definition for ErrorTier will be removed later

Reviewed By: iguridi

Differential Revision: D65688577

fbshipit-source-id: c96dfd69b3cee4f3bce86beb7f1714130e2fa3d0