roots: Add `cwd` to `InvocationRoots`
Summary:
I finally figured out what to do about the relative cwd issue - this diff proposes to store the cwd as a project relative path in the `InvocationRoots`. We can compute it there without hacks and that type is generally available in the right places

I considered also making a `WorkingDir` type that wraps this, but didn't feel necessary yet. We'll see

Reviewed By: perehonchuk

Differential Revision: D65845853

fbshipit-source-id: 9cba6c5f04761a5f9f2db954f965e7f879dcc6d1