Add prelude//:none
Summary: When writing selects for targets, it's useful to abstract that select under an alias. However, sometimes certain arms of the select (ex. no-san) are not expected to produce anything, which means normally we can't use alias for it. However, in these cases, we can still use an alias if we make the select point to a no-op target that produces nothing. This diff adds a `prelude//:none` target for this use case.

Reviewed By: perehonchuk

Differential Revision: D65715720

fbshipit-source-id: dcd54be652453c10dd69108bb82b087134a3fe8c