Move msvc `/nologo` to msvc_hermetic_exec.bat
Summary:
After the recent compiler argfile refactoring which results
in nested argfiles, cl.exe no longer respects /nologo. It seems to
only accept that if it is in the highest level argfile or directly
on the cli. So move it to msvc_hermetic_exec.bat which we use to wrap
all cl.exe invocations anywaym

Reviewed By: 8Keep

Differential Revision: D66119016

fbshipit-source-id: 054f68db3e2b2285b005e57dc19f38b97ed8c0ba