3/n: Update elp.bxl to improve dependency information
Summary:
One of the problems we have with the current ELP buck project model is it does not properly track includes per application, and relies on the "global include file hack" to add all include file paths it finds to a global list.

Not only is this inefficient, as it makes the search time for each include file longer through checking a larger list, it is also potentially inaccurate.

This is particularly relevant if we have a test that must run against prior versions of an API, in which case the include attribute will be the same each time, but it must resolve to a different file depending on the test target.

# This diff

For each target, walk through its dependencies, constructing a union of all the include paths it finds, so we can have accurate per-target include specifications.  The result is cached per target.

Also return the list of applications and dependencies relating to the target, so we can use these, rather than inferring them from the directory structure.

Reviewed By: perehonchuk

Differential Revision: D65941705

fbshipit-source-id: 54178deb822808fa282f3e27e2ac618f63c6a619