Improve formatting of incompatibility error output
Summary:
Four specific parts of this diff, I'll break them down

1. Pass in "alternate" flag so `IncompatiblePlatformReason` error display prints newlines after each dependency
2. Add a bit of extra whitespace to final output ("is incompatible with...") to make it look a bit prettier, per the suggestion in the linked task
3. (minor) update code pointer in comment to point to the right place (previous code pointer must be the wrong commit, or code moved or something). Intentionally not shortening the URL so it's clearer where the link is going to go and you can open it in your IDE more easily if you want.
4. Update tests for new output string(s)

NOTE: #4 (updating tests) is still a work in progress, I'm publishing this diff early for reviewing the other logic and will fix the tests before landing

## help needed:
Is this push safe; updating buck core code and www code (the Sandcastle tests) in the same diff? This error makes it seem like no
> fbs-validate-www-changes
> PUSH UNSAFE - Landing together does not mean they are deployed together. Changes to fbsource/www and fbsource in the same diff are currently an opt-in experience. If you understand the risks as outlined here https://www.internalfb.com/.../When_to_use_megarepo_0/ Please join this group (https://fb.workplace.com/groups/757968966024362) and rerun the job.

That said, this is just tests and also putting the test changes in a separate diff doesn't seem any better than doing it all in one diff?

Reviewed By: blackm00n

Differential Revision: D66050821

fbshipit-source-id: e288909af5489aaa2ec451b1ab19dcfd8748b406