22/n: elp.bxl: include prelude in targets to be searched
Summary:
The Erlang prelude should always be available in a BUCK erlang project.

Rather than require it to be specified explicitly in the `.elp.toml` file, always add it to the list of targets to be processed in `elp.bxl`

This causes 2 tests to fail because of listing details of all the prelude targets in the expected results, but we address this later in the stack when we only process prelude targets that are actually used in a project.

Reviewed By: perehonchuk

Differential Revision: D66304151

fbshipit-source-id: fa9db64504161646cec2919d4f041254597859ee