Create buck2.py utility script to enable multiplatform execution
Summary:
I created D66307563 which was a PowerShell script to build Buck2 on Windows, as the buck2.sh doesn't support it. However, cjhopman suggested the creation of a python script that we can use on all platforms and can be maintained better.

This diff intends to replicate the functionality of `buck2.sh` in order to iterate on local changes of the buck2 binary, while also taking care of Windows.

Reviewed By: JakobDegen

Differential Revision: D66397886

fbshipit-source-id: f673616727c658738bca642d4ffedf9f3ef8dd63