Filter out all non-source-owning targets for `arc pyre check`
Summary:
filteringfeatures

`arc pyre check` has a flag called `--report-unchecked-targets`. This target will report, say when the user invokes `arc pyre check fbcode//foo/bar/...`, which targets under `fbcode//foo/bar/...` are "unchecked".

The most frequent reason why a target is "unchecked" is that the target has set `typing=False` (directly on target or indirectly via package file) on it. However, another very common reason for unchecked target is that the target itself is a `python_binary()` or `python_test()` rule that doesn't own any sources. This second reason seems to be very uninteresting for most people, so it would be nice if we don't have to report them so `--report-unchecked-targets` can focus on reporting typing-disabled targets.

This diff tries to filter out all non-source-owning targets from Pyre's BXL script, by examining the `sources` attribute on the configured targets and discard the target if the attribute is empty.

Reviewed By: zsol

Differential Revision: D66525151

fbshipit-source-id: 4038389c695c679e44fe4fdd51f5dfbb60e5e101