Check code_coverage.file_path_filter on rust macros.
Summary:
Since finest granularity rustc can do coverage is crate (or target in buck world), we did not implement file_level filtering for rust coverage and we've always been using folder_path_filter every time we need to collect coverage for a rust file.

With this change we will be able to check if any of the srcs of a target is in the coverage ilter and then enable coverage for the whole target.

As result we will end up collecting less (unneeded) coverage and will be albe to simplify the CI setup. (rust will be handled just the same as c++ and python)

Reviewed By: fgasperij

Differential Revision: D66574505

fbshipit-source-id: cc17da0048d2e5e1e268373edb50e9ff34257271