buck2: Log instant event when creating new configurations
Summary:
We'd like to have a table that maps buck2 configuration full names (including the hash) to their constraint keys / values. To do this, we need all the configurations created during a build emitted in the log and sent to scribe so we can then send this data to downstream tables.

This uses the `LateBinding` machinery to make an `EventDispatcher` (implemented in buck2_events) available in buck2_core (which sits near the bottom of the dep stack, e.g. it can't take a direct dep on buck2_events). Only `InstantEvents` are exposed for now, can expose more later if we need them (and with a bit of tweaking of the interface to `EventDispatcher` because it has to be object-safe when used in a type-erased fashion).

This emits all information about the configuration:
* Name
* Constraint settings and values

Reviewed By: JakobDegen

Differential Revision: D66578831

fbshipit-source-id: 212f8f59826587dce32e00a8ffab3d4f2c2ba68b