Always use the java agent in the installer
Summary: The native agent is only used for API 19 and below, and in 2020 we added a config to always use the Java agent regardless of the API level (i.e. for API 19 and below too). That had no issues on buck1 (it was inadvertently turned off with buck2). Given it's been another 4 years, seems safe that we can just always use the Java agent.

Reviewed By: perehonchuk

Differential Revision: D66537790

fbshipit-source-id: dec1b3ea4c12b60b1bf3a155586f9dab655542cd