Fix incorrect path bug in planner script
Summary:
This is the GNU version of D66563629. Quickly, we overwrite this path
variable when we don't actually intend to. Fortunately, even though we
include it in the plan, we never actually use it in the GNU implementation
so this change is a non-functional change.

If we started to it would be incorrect often, so fix it before that becomes an issue.

Reviewed By: jtbraun

Differential Revision: D66641653

fbshipit-source-id: cb8deeb75ad951443f3e8ea7bbc3abfbf95538e0