Add XCFrameworkInfo provider
Summary:
Not 100% sure this is a good idea, but when consuming the XCFramework in the SPM packager, I struggle to get the name of the module for each framework without hackery (e.g. assuming Foo.xcframework supplies Foo).

I've added a provider here; the downside is this would stop you from using prebuilt XCFrameworks in SPM packages as they wouldn't have this provider.

Reviewed By: milend

Differential Revision: D66536726

fbshipit-source-id: 2300bda66cdb71762075a816781348e540fac8ca