Copy BXL scripts from arvr/tools/vsgo/bxl to fbcode/buck2/prelude/ide_integrations/visual_studio
Summary:
So that external users of buck2 can benefit from this Visual Studio project generator.

## Context

VSGO is the Visual Studio project generator for buck targets and supported by Windows DevX. This half, we took the opportunity to rewrite and cleanup it using BXL which gives massive performance boost. See https://fb.workplace.com/groups/1496591530773407/posts/1898812170551339 Now that the rewrite is done and is the default implementation used internally, this diff moves the location to buck2/prelude to open-source the tool and in hope to benefit external buck2 users.

Differential Revision: D66523063

fbshipit-source-id: 7a421d872c02755ca4a10c4aed68e8b02fe0ab4d