buck2: Introduce `DepMetadata` and library versions
Summary:
Awhile ago we (AI Software Platform) added the `accelerator_toolkit_version` buildinfo entry to report which gpu toolkit version was used by this binary.

For a time, this was fine - we knew roughly how much in prod used specific versions of e.g. cuda.

However, this wasn't totally accurate - our datasets reported binaries using specific versions of cuda even if they didn't link against it!

We'd ideally like something that only reports a particular version of cuda _if it links against cuda_, meaning it's present on the link line - that is, the final link line includes one or more CUDA libraries.

This diff introduces a new record to keep track of "link metadata" which reports precisely this.

While adding a new provider might be attractive, the reality of multiple languages, link styles, preferences, and strategies, and general messiness of C++ linking mean that back-channeling this information through `LinkInfo` is the more pragmatic option.

For CUDA - which is our initial target for accurate version telemetry - we mostly care about C++ and Python, so this stack implements C++ and Python support.

Future support is likely a good idea - widely-distributed library authors (like in Core Data and Privacy) could make use of this mechanism to understand what versions of their libraries are running in production.

Reviewed By: IanChilds

Differential Revision: D66532546

fbshipit-source-id: 07ccc25b2f2acbb61cc2d3627ef5959e6d852a6a