Add extra logging to debug timeout issue
Summary:
# Context
We are seeing a strange issue in which some test suites runs timeout. When we look at the test_runner.log, we see that the test node seems to be started exactly 9m30s after execution starts, which coincides with a hard timeout we set and is preceded by the test_exec app being stopped.

# Problem
This has been impossible to reproduce locally. So far, we can rely on the timeout message to find samples in CI logs, download archived logs, etc.

# This diff
So we start by adding finer grained logs, to try to get more insights of where the execution is, and figure out what is blocking.

One thing we want to figure out is if control actually makes it to the spawn_port() for the test-node. To be sure, we manually include a timestamp in the log entry, in case the logger is somehow delaying that event and showing a bad timestamp

Reviewed By: acw224

Differential Revision: D66759101

fbshipit-source-id: 33ec96e99c9f590378afc0a17b9c046a33f0cf0c