Make execute_over_thrift field optional
Summary:
If `ClientCfg.execute_over_thrift` is a non-optional bool defaulting to false, we have no way to distinguish between an explicit specification of GRPC, and non-specification.

This is a problem if we want to use feature config to control the mode. In cases such as frecli and buck2, who have their own RPC mode control, we want to be able to respect their request even if the feature config says something else.

To fix this: make the field optional and do the defaulting inside REClient.

Differential Revision: D66702233

fbshipit-source-id: 83196ec5cf463d39b31c558db680e71c835afe7f