Removing the conditional for asic-grid
Summary: asic_platforms and asic_devinfra_xtensa_platform are used to direct asic RE jobs to the right platform. As we've been doing a slowroll migration of actions from asic-grid to asic-prod, we added a conditional so we could use configs in mode files to determine which RE platform should be used. Now that we're finished with the slow roll, we can move the remaining actions to asic-prod with this change.

Differential Revision: D66437071

fbshipit-source-id: 78747fd1b0fdc6d4e91b5e8c993404df6c3ee956