Track origin of external buckconfigs in proto efficiently
Summary:
Let's store the origin of external buckconfigs in the event for buck configs too.

Rather than storing the origin in ConfigValue which I have prototyped in D66704002  (which makes the events much harder to read), we keep all buckconfigs coming from a single file as `GlobalExternalConfig` which already has the origin.  I find this representation easier to read and more efficient.

Reviewed By: IanChilds

Differential Revision: D66650209

fbshipit-source-id: 6e371f754fb0ed3d55d7c8e382cb20440058c4ed