Store the origin of the outside of repo configs
Summary:
Curently, we have a single bmap for all the parsed config values for external files (tracked in `started_parse`). However, I want to store provenance information for where each config is coming from (for logging purposes), so we can't keep a single map. Instead, we keep a vector of them with origin information.

Before we invoke parsing of in-repo configs (i.e. in `finish_parse`), we make sure to combine the vector into a single map first so there is no disruption down the line.

Reviewed By: IanChilds

Differential Revision: D66542534

fbshipit-source-id: 046b9c274cb0059e7dea318e87c07dcaeaca4b02