simplify http "is this error retryable?" API
Summary:
Like it says in the title. Instead of requiring you to turn your error into a
HTTP error and having `http_retry` infer whether or not the error is retryable,
expose whether http errors are thought to be retryable, and expect the caller
to use that.

I plan to use this in the next diff.

Reviewed By: IanChilds

Differential Revision: D66882748

fbshipit-source-id: 5cbd9f588e73871a87406bae8937795187a2f4f0