Use ActionExecutionKind::LocalCache when we get a dep file hit without running dep files
Summary: This separates our logging for the case where we actually do a dep-file comparison, and the case where we just utilize dep-files to get a local cache hit (i.e. we don't actually do the dep-file comparison, we just compare all of the inputs of the action and verify they are the same).

Reviewed By: ezgicicek

Differential Revision: D66881618

fbshipit-source-id: 672336295143041d66c7d8e27f8326d6ec78eb6a