log memory usage from RE action execution metadata
Summary: We want to log action's max memory usage executed on RE the same way as we track it locally. RE sends as execution action details that contains memory stat. At the moment this stat is not available if an action comes from cache. I'll add it as soon as RE implements it (See T209911776)

Reviewed By: perehonchuk

Differential Revision: D66961391

fbshipit-source-id: b941ce635dcd62249896706f070e5feacdbce3fb