Log experiments as tag events to scuba
Summary:
Now that we have all the values from external buckconfigs, we can pull out the values of experiments from `.buckconfig.d/experiments_from_buck_start`  (which are stored in external conifgs) and log them as tag events with prefix `experiments.`.

One concern is the size of these experiments... We currently have ~50 experiments so for each command, we will be logging these.

By looking at idata, tags column has 19.6GB which is  0.2% of storage overall.  p75 of its size is 11. So by adding 50 more elements, we might increase storage usage 1% more, so approximately 3 days of retention. This is the worst case though since experiments are only added to user builds which are around 4% of total builds. So impact on retention will be negligible.

Reviewed By: scottcao

Differential Revision: D66882006

fbshipit-source-id: 53760cfd345c5a3762603fb20b4e87b74dc07917