pass XDG_RUNTIME_DIR even if systemd disabled
Summary: When we run e2e tests there could be a case when we first build buck having systemd disabled (and not passing `XDG_RUNTIME_DIR`) but then we run built buck in test that require systemd. Because we didn't have  `XDG_RUNTIME_DIR` in the first place we were not able to pass it to test. See D67025683 for example.

Reviewed By: blackm00n

Differential Revision: D67027979

fbshipit-source-id: dd87ca3863e50526139c894d9ad1cab82af7540f