don't include runtime files for inplace build step
Summary:
Update `_pex_modules_args` to only include manifest files in the command dependencies.

Update `_make_py_package_impl` to add dep artifacts to build command for standalone builds, and to runtime files for inplace builds.

This makes make_par - inplace no longer depend on the artifacts going into the build, only on the manifest files, which only change if the name of an artifact changes.

Reviewed By: cxxxs

Differential Revision: D66900084

fbshipit-source-id: f3246b6f13724e0fcb9e7b13b44b1fc55ceae2f1