Add how to cache error for building artifacts
Summary: This diff add how to doc for caching error for building artifacts. This may similar with how to handle errors docs, but build artifacts api is a special one which have some limitations and also need a one more step. So it is worth to have another doc to explain it

Reviewed By: IanChilds

Differential Revision: D67116890

fbshipit-source-id: cfbeba371faf98523c6360fbfd04383d2d1240c7