Introduce CellNameForConfigurationResolution
Summary:
I came across us passing around a cell name for exec platform
resolution and didn't know what it was for (and I probably wrote it). So
just adding this newtype with some explanation so that people can figure
it out without needing to dig through a bunch of code to figure it out.

Reviewed By: JakobDegen

Differential Revision: D67299812

fbshipit-source-id: a3c3bbebef328907c7a7c2ef76ff751fe885380f