Add get_representative_config_flags
Summary:
This extracts the "representative config flags" from the ExpandedArgv.
These flags are the -c/--config/--config-file flags from the command line,
except when those are from in-repo modefiles we include the modefile instead of
the flags. This gives us a more compact, more stable, and more recognizable form
of the flags than using the -c flags directly.

Reviewed By: IanChilds

Differential Revision: D67046990

fbshipit-source-id: 00df8ad1b1882ca44b1c669f3ecd09c405dbc7d3