Don't unwrap possibly canceled result
Summary:
These are written expecting the result to not be cancelled because
they know that they don't cancel the spawned task. But, it the result from those
calls could be canceled by the task itself computing a canceled result (for
example because they look up a canceled dep).

This changes these panics to propagate the cancelation instead.

Reviewed By: JakobDegen

Differential Revision: D65363689

fbshipit-source-id: 153e07dd7471f734fce19d0b9584b2c25ce56c86