Add CommandExecutionStatus::WorkerFailure
Summary:
In the top of the stack I'd like to distinguish between test executions which failed because the worker failed, and test executions which failed because the command failed. This diff adds a variant to CommandExecutionStatus to allow distinguishing a worker failure from a regular failure. It is currently not used for clarity. The top of the stack will actually instantiate this.

NOTE: This does not expose WorkerFailure to protobuf. I wasn't sure if this should be kept internal or not. The next diff exposes to protobuf, and I'll squash the two if needed.

Reviewed By: christolliday

Differential Revision: D67734576

fbshipit-source-id: 2399bcc4559339c1ee712c0b2a8ed39d52e04716