`from_str` to return buck2_error
Summary:
Have `from_str` for `AbsPathBuf` to return `buck2_error` instead of anyhow.

A few inputs to `clap` had to be converted to string as conversion could fail with `buck2_error` which clap doesn't understand. So instead take in a String instead and do the conversion in code.

This allows us to remove a bunch of conversions and dependency on `anyhow` from `buck2_core`

Reviewed By: JakobDegen

Differential Revision: D67764601

fbshipit-source-id: 83a46751eb80bdd677945e9f55b89735bd838ec9