Add listing to critical path tests
Summary: The tests were just checking for the existence in the reverse order with zip dropping the rest (expected was shorter length). For the next diff, we prepend  a new kind to the beginning so we need to ensure that the first item (listing) is there so we need to extend the expected with the right one.

Reviewed By: IanChilds

Differential Revision: D67858340

fbshipit-source-id: 872ee99e2a0bc45f853588474a7f15137f41c3e3