Add option to force include link group shared library even if no symbols were referenced
Summary:
Add an option to wrap the link group shared library with `--push-state --no-as-needed <link_group.so> --pop-state`. This is helpful when users want to leverage link groups to ship shared libraries that are not referenced via symbols.

It would be great if we can get rid of this hack and just use resources, but we currently do not have a language-agnostic resource feature that is transparent to users. Currently, resources are only implemented for C++ and Python, and Python requires service owners to initialize a resource manager.

Reviewed By: BrandonTheBuilder

Differential Revision: D67881936

fbshipit-source-id: c76b1cfa5ac858f2926edc274676d5ab4542fb05