Add TimeSpentSynchronizingAndWaiting to critical path
Summary:
We aim to improve our accounting of critical path and one of the missing pieces is "the time spent synchronizing changes and waiting for concurrent commands to finish" (see attached task).

This is the time between `CommandStart` and `CommandCriticalStart`. Let's log this as the first entry in the critical path with a unique kind `"time-spent-synchronizing-and-waiting"`.

To store this in critical path, we have to

1.  pass around the command start time to dice ctx so that  we can compute the elapsed time when we fire the  `CommandCriticalStart`.
2. pass this elapsed time further to `BuildSignalsContext` so that we prepend it to the critical path.

An alternative would be to parse all the events (not sure this is possible) at critical path calculation to compute the time between  `CommandStart` and `CommandCriticalStart` but this would be an overhead.

TODO: Need to update Buck UI to add tis new kind in its graphql query + rendering

Reviewed By: cjhopman

Differential Revision: D67794903

fbshipit-source-id: 7da9de3adca9a6d38912f8c8f3bb1ef5a710d2ef