Put eval anon target in scope spawn
Summary:
As comment (same comment for eval bxl) said, we would use block_in_place in eval, it will prevent the inner future from being polled and yielded. So, for cancellation observers to work properly within the dice cancellable future context, we need the future that it's attached to the cancellation context can yield and be polled. To ensure that, we have to spawn the future that then enters block_in_place.

This diff put the eval bxl in scope spawn

Reviewed By: JakobDegen

Differential Revision: D67992459

fbshipit-source-id: 4837d065afab22f0af3f2d5264ed5fd748f1892f