Require error tag for buck2_error! macro
Summary:
Last part of the code where `buck2_error` can be created.

Have `buck2_error!` macro take in an ErrorTag instead of an array, making it possible to enforce an ErrorTag (Since ErrorTag can't be set as empty but we can't enforce a non-empty array at compile time unless there's something I'm not aware of).

All usage of this macro have at most 1 ErrorTag anyways so this doesn't make a difference at the moment, and tags can be added as an array through the `.tag()` function

Reviewed By: JakobDegen

Differential Revision: D67998689

fbshipit-source-id: 397fbc272f05d97713ec66284c563035e516e46a