Require tag for all buck2_error_derive enum types
Summary:
Force an ErrorTag for all current and future enum types that are derived from buck2_error_derive. Can be as simple as `Input`, `Environment`, and `Tier0`.

Note that I tagged all the current ones to the best of my knowledge and could be wrong. Please fix or let me know anyone finds anything that seems wrong

Reviewed By: JakobDegen

Differential Revision: D67954001

fbshipit-source-id: 17798faa4d7365e1d24c3eddac2bcc55d1e66b95