Streaming worker tools
Summary: Worker tools with tons (millions) of actions queued up will have millions of async tasks consuming memory. This can reach the order of gigabytes. To avoid this, workers who's execution model is inherently async can opt to use streaming execution. This keeps a single async stack for the all actions. See next diff for an example. I made this opt in because writing workers in this way is complicated and usually not desirable.

Reviewed By: christolliday

Differential Revision: D67939003

fbshipit-source-id: eb179718221370c0d8e4ebfbf2d0e1b38a38e8a7