Sort errors in invocation record so that errors[0] == best_error
Summary:
Currently the invocation record contains a list of errors and a subset of fields for the 'best error', but not the error message for the best error, which I want to use in the next diff.

We could add best_error_message as a new field, but if we use `errors[0].message` we reduce data duplication and improve the buck2 UI since the error list is ranked by severity.
In theory the existing ordering is based on when the errors were hit, but that ordering is not guaranteed and doesn't apply to client vs action errors. This should preserve the existing ordering for equally ranked action errors.

Reviewed By: iguridi

Differential Revision: D68462352

fbshipit-source-id: 6db16f6a1d8486bcc59d84ddc117b5d12b492e5b