[ct_error_printer} Do not format exception when expecting an exception.
Summary:
When one uses assertException(..., ...), the "received exeption" was printer using erl_error:format_exception(....), whereas the "expected exception" was not.

This lead to difference in format that made them hard to compare.

In this diff, we print resulted exception using the usual "~p", to make it comparable to the expected one.

Reviewed By: alanz

Differential Revision: D68023395

fbshipit-source-id: e40df50bfb5096fa88712bc5bc31a03ab13cc2ff