Allow logging tagged errors from installer
Summary:
Android DevX is interested in improving installer errors using error tags and categories. This lets the installer send tags and categories that get added to a buck2_error and logged to buck2 telemetry.

We want to support string tags for this use case since syncing error.proto changes to the installer is not simple and adds friction to adding external tags.
Currently these won't log to buck2 error_tags they just show up in the category key since they are strings, but I want to change 'context_for_key' string values to be logged to error_tags anyway.

Reviewed By: navidqar

Differential Revision: D68452760

fbshipit-source-id: 460f32ca18f1fcc3a7d03e095c2ab7d0a81ecd84