Use struct for source location
Summary:
Storing source_location in buck2_error as a struct that can be formatted, instead of an `Option<String>`.

Motivation is to use the type name (currently stored as extra) separately without parsing this string.

Also making it non-optional since the only case where source_location is None is if the path is outside of buck2/app. I expect that we shouldn't have paths like that since buck2_error isn't used externally, but if we do we still want to know what the path is. Added an "external:" prefix so they can be identified if needed.

Reviewed By: Will-MingLun-Li

Differential Revision: D68351423

fbshipit-source-id: 84cfb345f143b9f71a193bff3bb16f3e40b4e980