Don't include forward nodes when building with a target-universe
Summary: Right now, when you have a transition and you try to do a `buck2 build` or a `buck2 run` with `--target-universe`, the command will try to build two nodes - the "forward" node and the actual node. For `build`, it means we report the same target as building twice (when you use `--show-output`, for `run` it means that the run completely fails because there are two build targets.

Reviewed By: JakobDegen

Differential Revision: D68554360

fbshipit-source-id: c5d6dedfa0cdf077ab0a010a3d41db787b183dfd