Exclude source location from category key if type name is present
Summary:
We should shorten category keys in places where it is possible without reducing uniqueness of keys, especially if they include things that are likely to change with refactoring.

We should keep type names if we have them though, especially for enum variants since many of those don't have unique tags.

For now keeping source locations in the key if there isn't a type name, but we can also remove those later for errors where we can ensure tags are sufficient as a unique key.

Reviewed By: Will-MingLun-Li

Differential Revision: D68519035

fbshipit-source-id: 8f644b3a9aa41b8a50a5a84e6c8ef6a403d83721