Log buck2 test failures to scuba as Input error
Summary:
Buck2 test failures (The test executes successfully, but the test itself fails) are currently only logged via stderr/stdout.

In Scuba, these gets marked as failed, but there are no information about the failure (No error message of any kind and no error tags), and it just gets displayed as `null` which is confusing and wrong for error categorization.

Just add an error message and mark these as `Input`, we could easily include the exact test failures if we want but I'm not sure how valuable those data would be as there could be hundreds of regular test failures

Reviewed By: JakobDegen

Differential Revision: D68601386

fbshipit-source-id: c6bed97bd537a09d212205112979a7cc7def5abe