Add tag for error from Tpx binary
Summary:
Vast majority of infra errors for `buck2 test` are coming from TPX binary, so adding a tag for it to make it more distinguishable for now.

Skimming through Scuba, most of these are actually user errors for passing in the wrong argument. This is mostly a note for myself but we could do a string match for `Error parsing test runner arguments` to mark 90% of these are User Errors, but that's a brittle solution especially since that's just an error context and not even an actual ErrorType.

Reviewed By: JakobDegen

Differential Revision: D68646142

fbshipit-source-id: 2fb44f175d43eea01b775b8207fb7c5f7f14ad7e