RFC: Have buck provide the test config to tpx 1/3
Summary:
This is a split diff.

The goal of the whole stack is to pass enough info to tpx that it can populate the reproduction command/instructions on testx.  That information is the host that is running buck (e.g. "what should I be running this on my mac or my windows vm or my linux box?"), and the configuration flags passed to buck (e.g. mode files, -c configs).

In this phase, our goal is just to consistently get the right flags both locally and in CI and log them to scuba.  The CI portion is already done and is logging now.  This is the local portion.

The next phase will be changing the repro commands/instructions and the entries in the test db to reflect these values.  That will happen once we validate that this code (and the CI code) is working as intended.

This diff is one of three:
1. (this diff) Makes the interpreter host platform public enough that I can access it from this code, and sends it in simplified string form to TPX.
2. Adds the "representative_config_flags" entry to the protobuf def of ClientContext so we can access it in the server, and passes the config slice of that to tpx
3. Passes the flagfile slice of representative_config_flags to tpx.  This is the smallest diff in the stack.

There have been several rounds of review on this stack - I am splitting it so that we can have parallel discussions (and because the latest round of discussion required larger changes to buck's codebase, so its easier to review in different diffs).

Reviewed By: JakobDegen

Differential Revision: D64497906

fbshipit-source-id: c1893c5b60e9f42fb5322520d35f985d2d6dcaf5