Have buck provide the test config to tpx 2/3
Summary:
See parent diff for full story.

This diff forwards the representative flags from the client to the server so it can be passed to tpx so tpx can use it for the repro  commands for the test runs.

Long-term, we should probably consolidate the different ways args are passed (sanitized_args, flagfiles, and representative_config_flags) - they all serve different purposes but have overlap.  Its a bit beyong my knowledge about how best to do that.

Reviewed By: JakobDegen

Differential Revision: D68515415

fbshipit-source-id: 8cae6c7739ec3933679bdd9019a01d9d1303b168