Forward max memory used from buck to tpx
Summary:
This is a really large diff that is just passing a value from buck's execution info to TPX's test result info.

If there is a better way, please let me know.

It might also make sense to use a proper struct to allow us to pass arbitrary stats - right now we hard code duration (and memory, now).  In the long-term, we may care about network or cpu, though, to be fair, only a handful of teams seem to care about those numbers today.

Reviewed By: samkevich

Differential Revision: D68564257

fbshipit-source-id: 44dfd8a1b96b618721b337e0a8cde1e5700259d9