enable delegation in action scope scgroup
Summary:
Apparently, there are cases when cgroups are created inside a test. Currently, tests, like any action, run in a cgroup, making it impossible to run a cgroup inside another cgroup.
This diff fixes this issue by allowing the creation of a cgroup inside an action's cgroup.
There might be issues with deleting actions cgroup if a test doesn't cleanup all resources. But this issue exists regardless of this diff, so I'm going to address it separately.

Reviewed By: Ben0mega, rajneesh

Differential Revision: D68219577

fbshipit-source-id: 82c46353a06d3a577e40bb569ab1d897f0c2fca0