Remove unwrap in streaming worker
Summary: See attached task. If response is Err then the next element on the stream should be a None and the stream will terminate anyway, this is not an issue for Runway jobs because we only use a single worker. But some jobs might want to recover from this case.

Reviewed By: DiZy

Differential Revision: D68667189

fbshipit-source-id: ba01d76b6cac40f1c3d4b27cb66fe74e3e8af5f5