Fix missing tags for memory pressure and disk space
Summary:
I realized that we only tag memory pressure + disk space warnings if they still occur on the final snapshot. However, this was not intended to be the case as the comments indicates. We should tag a build even if we show a warning that ends up disappearing at the end of the build.

To fix this, we can look at the peak memory/disk space which stores the maximum.

Reviewed By: IanChilds, rajneesh

Differential Revision: D68756813

fbshipit-source-id: 80e132f450e329a4d6ded2ec3d676a4541788270