Upload watchman and eden debug logs
Summary:
Upload `watchmanctl rage` and `eden debug` logs to our manifold bucket to help with debugging WatchmanServerErrors.

Note that this does regress the user experience when encountering this error. Measuring on my devserver, this added ~7s of delay if both are uploaded successfully. However, since this will only run when it encounters the error, having this regression now to eventually prevent the error from happening in the first place seems like a fine tradoff

Reviewed By: christolliday

Differential Revision: D68691885

fbshipit-source-id: e0a5c9ae99b95d371ede3c9b496d8c0728b2f16d