Add more context for Watchman/Eden rage log message
Summary:
Looked at Scuba to see if Watchman/Eden rage log upload is working as expected. Realized that I didn't add any contexts other than the link to logs.

I know what the links are since I added them, but it's probably not obvious to other people, so clarifying it a bit

Reviewed By: samkevich

Differential Revision: D68843933

fbshipit-source-id: 5c550a2b5346c5a235e48ef225bad59db7c976df