Add `ThinBoxSliceFrozenValue`
Summary:
Semantically a `Box<[FrozenValueTyped]>` that is better optimized than even `ThinBoxSlice` in the length one case. Explained in the diff.

Implementing this in a reasonable way required making a copy of the `ThinBoxSlice` impl. With more compromises I could have reused the code, but that would have lead to weird non-local invariants, so I think this is a much lesser evil.

Reviewed By: JakobDegen

Differential Revision: D66831083

fbshipit-source-id: 3b218a6ba5f45751ae937f9464d5b15f155b438d