Process external_configs after synchronization
Summary:
If there are file changes in between two identical commands, external config values were recorded in the event log twice. This is because  events were fired in dice updater which is called by the concurrency manager in a loop in order to get the updates in cases that files on disk have changed between commands.

This diff pulls out code to outside of this loop  so that these events are fired once we acquire access to DICE transaction. This way we only record them once, not each time we check whether the state is changed in between.

This diff also resolves another issue with picking up the correct external configs if `--reuse-current-config` is set. In that case, we reuse the external config data from the previous command, so we should record that.

Reviewed By: JakobDegen

Differential Revision: D68828945

fbshipit-source-id: f3059d08af09dc3562d786d6e3ad38ef664d7adb