Read daemon start up errors from file on connection failure
Summary:
Currently if daemon connection fails from the client we try to read stderr from the daemon, parse it to extract tags and show it to the user.

This has a few problems:
- Produces hard to read error messages without useful tag information.
- Because this error is ugly we have a hack to connect from the daemon even when initialization fails just to show a better error, that leaves the daemon in a bad state and requires running `buck2 kill`.
- Normal errors in daemon start up often won't write anything to stderr.
- Even if the daemon does write to stderr, it may not have finished writing to it by the time the client reads it.

This writes the error as serialized ErrorReport, so it can be deserialized along with tags on the client.

This adds a 500ms delay if the client fails to connect and the daemon doesn't log an error file. With no delay the error file hasn't been written by the daemon by the time the client tries to read it after a connection failure, at least for a failure caused by `BUCK2_TEST_DAEMON_RUN_ERROR=1`.
The timeout can probably be shorter but users should not be hitting this case often, plus if the daemon has crashed it also may need time to write to stderr. Not having a delay may account for some of the instances where connection fails but daemon stderr is empty.

Reviewed By: iguridi

Differential Revision: D68652392

fbshipit-source-id: cf6d15909b500dee0ea9e35b81ccae3e4f5c31c5