Test for TSet deduplication
Summary:
Was investigation some duplicated values in Android rules, and wondered if it was a TSet bug (and actually wrote the opposite test), but I think clearly the TSet behavior is correct (and it was a bug in the rules).

Not sure how much value the test has, but would break a lot of stuff if this changed so I think it's reasonable to have?

Reviewed By: christolliday

Differential Revision: D69050758

fbshipit-source-id: c49a8e5f7f0ad01ad0dc04077a853982c302884b