Update test showing bug in re config override
Summary:
Updating tests to show that  re config overrides that are not coming from config flags end up setting `new_configs_used=1` (via `new_config_indicator_only`). However, because `ConfigDiffTracker` is completely broken, this was not tested properly in D64477014, tests relying on diff_count which is always 0.

The issue is that when files as inputs to configs change, buck spends some time to check DICE leafs invoking the following code

https://www.internalfb.com/code/fbsource/[3968cda2ef0ffb172ba02d034a7edaa45208c367]/fbcode/buck2/app/buck2_common/src/legacy_configs/dice.rs?lines=186-200

However, if we filter re overrides, we still end up sending a signal that configs might have recomputed even though we don't propagate these config changes up the DICE graph. We should maybe fix this (later) but then we are paying the cost of it when changes are in the files not in commandline.

Reviewed By: blackm00n

Differential Revision: D69052026

fbshipit-source-id: 7b4407f448b36cfa11540704cc06772b2fdc0867