Remove ConfigDiffTracker
Summary:
See D68954545 for explanation of why `ConfigDiffTracker` is broken. Only part that works is `new_config_indicator_only`. This diff replaces it with a new event `CellHasNewConfigs` that is only fired when configs are computed in DICE.

Let's make the code reflect what is happening and remove code that is misleading and non-functional.

**Why not fix diff size tracker?**

Well, it is not straightforward to fix this in DICE and also these numbers are not really helpful. For instance, if I change modes, we will end up diff_count differnce in thousands. What do I do with it? What is actually helpful is to know that modefiles or some external configs have changed.

In the next diffs, I will be replacing it with a mechanism implemented outside of DICE to track actual values of external config diffs. That will be much better.

Reviewed By: IanChilds

Differential Revision: D69052025

fbshipit-source-id: 99ffbbf7b82729b1860a689f522f7b3a635aacbe